refactor: useTextToSpeech hook
